sync

+++ app/Controllers/Sadmin/Configs.php
@@ -210,15 +210,14 @@ class Configs extends Sadmin
             // print_r($data);
 
             // xử lý với favicon nếu có
-            if (isset($data['web_favicon']) && $data['web_favicon'] != '' && is_file(PUBLIC_PUBLIC_PATH . $data['web_favicon'])) {
+            if (isset($data['web_favicon']) && $data['web_favicon'] != '' && strpos($data['web_favicon'], '//') === false && is_file(PUBLIC_PUBLIC_PATH . $data['web_favicon'])) {
+                // nếu là file .PNG
                 if (strpos(strtolower($data['web_favicon']), '.png') !== false) {
                     echo PUBLIC_PUBLIC_PATH . $data['web_favicon'] . '<br>' . PHP_EOL;
 
-                    // 
-                    if (copy(PUBLIC_PUBLIC_PATH . $data['web_favicon'], PUBLIC_HTML_PATH . 'favicon.png')) {
-                        if (copy(PUBLIC_HTML_PATH . 'favicon.png', PUBLIC_PUBLIC_PATH . 'favicon.png')) {
-                            $data['web_favicon'] = 'favicon.png';
-                        }
+                    // copy 1 bản sao của file sang thư mục public
+                    if (copy(PUBLIC_PUBLIC_PATH . $data['web_favicon'], PUBLIC_PUBLIC_PATH . 'favicon.png')) {
+                        $data['web_favicon'] = 'favicon.png';
 
                         /**
                          * Chuyển đổi file png sang ico
@@ -227,20 +226,15 @@ class Configs extends Sadmin
                         include PUBLIC_HTML_PATH . 'vendor/php-ico/class-php-ico.php';
 
                         // 
-                        $ico_lib = new \PHP_ICO(PUBLIC_HTML_PATH . 'favicon.png');
-                        $ico_lib->save_ico(PUBLIC_HTML_PATH . 'favicon.ico');
-
-                        // 
-                        copy(PUBLIC_HTML_PATH . 'favicon.ico', PUBLIC_PUBLIC_PATH . 'favicon.ico');
+                        $ico_lib = new \PHP_ICO(PUBLIC_PUBLIC_PATH . 'favicon.png');
+                        $ico_lib->save_ico(PUBLIC_PUBLIC_PATH . 'favicon.ico');
                     }
                 } else if (strpos(strtolower($data['web_favicon']), '.ico') !== false) {
                     echo PUBLIC_PUBLIC_PATH . $data['web_favicon'] . '<br>' . PHP_EOL;
 
                     // 
-                    if (copy(PUBLIC_PUBLIC_PATH . $data['web_favicon'], PUBLIC_HTML_PATH . 'favicon.ico')) {
-                        if (copy(PUBLIC_HTML_PATH . 'favicon.ico', PUBLIC_PUBLIC_PATH . 'favicon.ico')) {
-                            $data['web_favicon'] = 'favicon.ico';
-                        }
+                    if (copy(PUBLIC_PUBLIC_PATH . $data['web_favicon'], PUBLIC_PUBLIC_PATH . 'favicon.ico')) {
+                        $data['web_favicon'] = 'favicon.ico';
                     }
                 }
             }
